解决一个类型不匹配的warning和coz的手机上页面弹出时keyboardWillShowNotification的bug

+++ bupt_bbs/bupt/ThemeViewController.m
@@ -124,7 +124,7 @@ static const int kNumOfPageToCache=5;
     return  NO;
 }
 -(void)tableView:(UITableView *)tableView willDisplayCell:(nonnull UITableViewCell *)cell forRowAtIndexPath:(nonnull NSIndexPath *)indexPath{
-    _page_cur_count=_pageRange.location+indexPath.row/_item_page_count;
+    _page_cur_count=(int)(_pageRange.location+indexPath.row/_item_page_count);
 }
 #pragma mark - 根据刷新方式刷新页面的内容
 -(void)pullDownToRefresh{
